# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AR Demo and candidate API
  version: 0.2.0
  description: 'Demo: RuoYi + PostgreSQL, fixed-role administrators and scenes. Docker
    Compose and local ArtifactStorage infrastructure are available without artifact
    HTTP endpoints. Historical cloud/day-night resource and publication contracts
    below await a real deliverable and client loading agreement; they are not implemented.'
servers:
- url: https://api.example.invalid
  description: placeholder
paths:
...
  /system/user/list:
    get:
      operationId: listUsers
      summary: 账号分页列表
      description: 账号分页列表。沿用若依 code/msg 响应；业务错误可能使用 HTTP 200。
      security:
      - bearerAuth: []
      parameters:
      - name: pageNum
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
      x-implementation: implemented
      tags:
      - Demo 已实现
  /system/user:
    post:
      operationId: createUser
      summary: 创建固定角色管理员
      description: 创建固定角色管理员。沿用若依 code/msg 响应；业务错误可能使用 HTTP 200。
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
      x-implementation: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      tags:
      - Demo 已实现
  /system/user/changeStatus:
    put:
      operationId: changeUserStatus
      summary: 修改账号状态
      description: 修改账号状态。沿用若依 code/msg 响应；业务错误可能使用 HTTP 200。
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
      x-implementation: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountStatus'
      tags:
      - Demo 已实现
  /system/user/resetPwd:
    put:
      operationId: resetPassword
      summary: 重置密码并撤销旧凭证
      description: 重置密码并撤销旧凭证。沿用若依 code/msg 响应；业务错误可能使用 HTTP 200。
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkResponse'
      x-implementation: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountPassword'
      tags:
      - Demo 已实现
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FrameworkResponse:
      type: object
      additionalProperties: true
      properties:
        code:
          type: integer
        msg:
          type: string
    AccountCreate:
      type: object
      additionalProperties: false
      properties:
        userName:
          type: string
          pattern: ^[A-Za-z0-9_]{3,30}$
        nickName:
          type: string
          minLength: 1
          maxLength: 30
        password:
          type: string
          minLength: 12
          maxLength: 64
      required:
      - userName
      - nickName
      - password
    AccountStatus:
      type: object
      additionalProperties: false
      properties:
        userId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - '0'
          - '1'
      required:
      - userId
      - status
    AccountPassword:
      type: object
      additionalProperties: false
      properties:
        userId:
          type: integer
          format: int64
        password:
          type: string
          minLength: 12
          maxLength: 64
      required:
      - userId
      - password
...